fix: review-findings hardening + st<->rg oracle divergences
Working-tree review-findings pass (Fable Edition) plus two oracle-caught
st<->rg divergences surfaced while validating the branch:

- index/encoding.rs: decode a truncated UTF-16 trailing byte (odd byte
  count after the BOM) as U+FFFD instead of dropping it, matching the
  WHATWG decoder (and ripgrep). Dropping it let `-x parse` match st's
  "parse" but not rg's "parse\u{FFFD}" (oracle fixture
  repro_e1c1603c26349124).
- search/mod.rs + cli/render/mod.rs: compile the verifier and output
  regexes with crlf(true), matching the oracle's `rg --crlf`. A trailing
  `\r` at end-of-line is now a terminator, so `-x pat` matches a final
  line "pat\r" and submatch extraction stays consistent with the match
  decision (oracle fixture repro_e1477df13c5a98f4).

Also includes the batched review-findings perf/security work and the
release-notes automation (docs/RELEASE.md, scripts/extract_release_notes.py,
release.yml body_path) and a stable-toolchain pin (rust-toolchain.toml).
CHANGELOG updated.

Verified: cargo fmt, cargo clippy -D warnings, cargo test, cargo test
--features symbols, and oracle_self + oracle_cli at PROPTEST_CASES=512 all
green. Both committed oracle fixtures pass.

+++ src/index/tests.rs
@@ -971,11 +971,10 @@ fn update_from_git_classifies_dangling_symlink_as_change_not_delete() {
     drop(index);
 }
 
-/// Bug 10 regression: a *bounded* `update_from_git` (the search hot path) must
-/// never do a synchronous full rebuild when the overlay would exceed its cap.
-/// It returns `OverlayFull` (so the caller searches stale and spawns the async
-/// catch-up) instead of blocking the latency budget on a `build_index`. The
-/// *unbounded* call (`st update`) still rebuilds inline.
+/// A *bounded* `update_from_git` (the search hot path) must never do a synchronous
+/// full rebuild when the overlay exceeds its cap. It returns `OverlayFull` to spawn
+/// an async catch-up instead of blocking search latency. The *unbounded* call
+/// (`st update`) still rebuilds inline.
 #[test]
 fn bounded_update_returns_overlay_full_without_inline_rebuild() {
     use crate::index::freshness::{UpdateLimits, UpdateOutcome};
@@ -1000,7 +999,7 @@ fn bounded_update_returns_overlay_full_without_inline_rebuild() {
 
     // Enough untracked files that the overlay would exceed 50% of the base
     // (build also indexes the repo's own tracked files), tripping OverlayFull.
-    let n_new = (base_docs as usize) + 4;
+    let n_new = base_docs + 4;
     for i in 0..n_new {
         std::fs::write(
             repo.path().join(format!("added_{i:04}.rs")),
